Add broken setup.py
This script will allow the package to be installed, but not run.

It will fail to run because some parts of setuptools do not appear to work with
the package_dir option, and the package's name is not the name of its source
directory.

If that worked, it would fail later, because the resources are not included in
the package.